+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Autoscaler Change Log
 0.4.0 under development
 ------------------------
 
+- New estimator 'StaticWorkflow' for Token-like predictions
 
 0.3.0
 -----------------------------